Break up parser into multiple files
