--- Broccolini/Parsing/NodeType.cs
@@ -1,10 +1,0 @@
-namespace Broccolini.Parsing;
-
-internal enum NodeType
-{
-    Section,
-    Comment,
-    KeyValue,
-    Unrecognized,
-    Epsilon,
-}
