--- Broccolini/Parsing/TriviaParseContext.cs
@@ -1,7 +1,0 @@
-namespace Broccolini.Parsing;
-
-internal enum TriviaParseContext
-{
-    TopLevelSectionChild,
-    SectionChild,
-}
